requirements.txt
astroid==2.0.1
confusable-homoglyphs==3.1.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==1.11
django-bootstrap4==0.0.6
django-friendship==1.8.0
django-registration==2.4.1
django-tinymce==2.7.0
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pylint==2.0.0
python-decouple==3.1
pytz==2018.5
six==1.11.0
static3==0.7.0
typed-ast==1.1.0
unicodecsv==0.14.1
Unipath==1.1
whitenoise==3.3.1
wrapt==1.10.11
